+++ server/src/main/java/org/opensearch/index/query/QueryCoordinatorContext.java
@@ -8,14 +8,18 @@
 
 package org.opensearch.index.query;
 
+import org.opensearch.action.IndicesRequest;
 import org.opensearch.common.annotation.PublicApi;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
+import org.opensearch.index.IndexService;
 import org.opensearch.search.pipeline.PipelinedRequest;
 import org.opensearch.transport.client.Client;
 
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.function.BiConsumer;
 
@@ -31,11 +35,17 @@ import java.util.function.BiConsumer;
 @PublicApi(since = "2.19.0")
 public class QueryCoordinatorContext implements QueryRewriteContext {
     private final QueryRewriteContext rewriteContext;
-    private final PipelinedRequest searchRequest;
+    private final IndicesRequest searchRequest;
+    private final List<IndexService> targetIndexServices;
 
-    public QueryCoordinatorContext(QueryRewriteContext rewriteContext, PipelinedRequest searchRequest) {
+    public QueryCoordinatorContext(
+        QueryRewriteContext rewriteContext,
+        IndicesRequest searchRequest,
+        List<IndexService> targetIndexServices
+    ) {
         this.rewriteContext = rewriteContext;
         this.searchRequest = searchRequest;
+        this.targetIndexServices = targetIndexServices;
     }
 
     @Override
@@ -84,10 +94,14 @@ public class QueryCoordinatorContext implements QueryRewriteContext {
     }
 
     public Map<String, Object> getContextVariables() {
+        if (searchRequest instanceof PipelinedRequest) {
+            return new HashMap<>(((PipelinedRequest) searchRequest).getPipelineProcessingContext().getAttributes());
+        } else {
+            return Collections.emptyMap();
+        }
+    }
 
-        // Read from pipeline context
-        Map<String, Object> contextVariables = new HashMap<>(searchRequest.getPipelineProcessingContext().getAttributes());
-
-        return contextVariables;
+    public List<IndexService> getTargetIndexServices() {
+        return Collections.unmodifiableList(targetIndexServices);
     }
 }